test: move utils into their own package
