+++ src/test/scala/chiseltest/simulator/Icarus.scala
@@ -2,6 +2,7 @@
 
 package chiseltest.simulator
 
+import chiseltest.utils.CaptureStdout
 import org.scalatest.flatspec.AnyFlatSpec
